Make scope bar buttons AXCheckbox'es for accessibility
This notifies the user the choice is exclusive to only one
button. Also including these buttons in an AXRadioGroup adds
positional indication to the radio buttons (like "2 of 3")
which gives sense of how many values there are to choose from,
where in the list of those the user currently is, and, if there
are more sequential AXRadioGroups (which is not the case here),
it distinguishes in which one the user is making a choice.