Add “find bundle for new items” API
We need this when installing standalone bundle items, saving macros, and when creating items in the bundle editor without a bundle selected.

The user experience of this code is however rather poor in that you can’t actually crete new bundles, the dialog text is not specific, and the logic for what bundle to preselect can be improved (prefer previously created bundle falling back on a local bundle, if any).