Only skip setting search folder for ⇧⌘F if already “owner”
The user may use ⇧⌘F either to bring existing results to front, or to start a new folder search for a potential different project.

If the dialog is not visible, we always initialize it fully. If already visible, we previously would not setup search folder since the current search folder might have been manually set by the user, which we do not want to overwrite.