Let file chooser read ‘followSymbolicLinks’ from .tm_properties
Set this to ‘true’ if you want the file chooser (⌘T) to follow symbolic links.

The reason for making it a ‘.tm_properties’ setting is only because I am not yet sure how to best integrate this into the UI. The find dialog has an identical UI option which is global option — it would make sense to coalesce these two options, and while doing that, make the find dialog’s version a project/path setting rather than have it be global.

Closes #984.