Move some Go menu items to Navigate menu
The Go to Symbol and Line are now in the Navigate menu, which means everything in the Go menu is about selecting a file or folder (for the file browser location).

The Navigate menu is about navigating around in the current file.

Also change menu item title “Go to Counterpart…” → “Go to Related File”