Keep last effective tab sticky when closing tabs
When the last effective tab is sticky, closing a tab (after switching to
another tab) would cause the next tab in the overflow menu to become sticky.

This also addresses a couple of cases where the effective tab would not be
rendered (e.g., closing a tab when the last tab in the overflow menu is sticky).