Reload file using most probable charset when it changes on disk
Previously if the file had been loaded as ASCII/UTF-8 and was changed on disk to something other than UTF-8, TextMate would go into infinite loop (retrying with UTF-8). It now uses the encoding classifier to find the most probable encoding (based on learned frequencies) falling back on ISO-8859-1, if there is no candidate.

Ideally it would bring up a warning, informing the user of the problem and allowing a manual correction of character set.

Fixes #1073.