Add EOF invisible
I thought this was nice just because, in the very specific case that
a file ends without a newline and invisiblesMap is set to not show
spaces, you couldn't tell whether or not the file ends with spaces.

The EOF character defaults to what is (according to wikipedia) a
standard unicode character for it, and may be chaned in invisiblesMap
by using the 'D' character (you know, as in '^D').