Allow scmStatus to be explicitly enabled
Previously, scmStatus would not be enabled on remote volumes
even if, say, a .tm_properties glob explictly matched a remote
volume and set scmStatus = true. This patch allows setting
scmStatus = enable to skip the local-volume check. Backwards
compatible with the boolean scmStatus setting.

Also defined are scmStatus = disable, local, rootvol.
disable is the same as false, local is the same as true (and is
the default as before), rootvol only enables files on the "/"
volume.