Allow scmStatus to be explicitly enabled
Previously, scmStatus would not be enabled on remote volumes even if,
say, a .tm_properties setting explicitly matched a remote volume and
set scmStatus = true. This patch allows setting scmStatus = everyDisk
to skip the local-volume check.

Also defined are scmStatus = localDisk and systemDisk. localDisk is
the same as true (and is the default as before), systemDisk only
enables files on the "/" volume.