Embryo of developer documentation :-)
This is a document with information I found most useful as a
new contributor to TextMate - basic stuff a new contributor is likely
to get in contact with.

It's of course incomplete, sparse, can be not well worded and even contain
mistakes. But someone had to make the first shot :-)